Create 471 - Encode String with Shortest Length
Given a non-empty string, encode the string such that its encoded length is the shortest.

The encoding rule is:, k[encoded_string]where the encoded_string inside the square brackets is being repeated exactly k times.

Note:
k will be a positive integer and encoded string will not be empty or have extra space.
You may assume that the input string contains only lowercase English letters. The string's length is at most 160.
If an encoding process does not make the string shorter, then do not encode it. If there are several solutions, return any of them is fine.
Example 1:

Input: "aaa"
Output: "aaa"
Explanation: There is no way to encode it such that it is shorter than the input string, so we do not encode it.
Example 2:

Input: "aaaaa"
Output: "5[a]"
Explanation: "5[a]" is shorter than "aaaaa" by 1 character.
Example 3:

Input: "aaaaaaaaaa"
Output: "10[a]"
Explanation: "a9[a]" or "9[a]a" are also valid solutions, both of them have the same length = 5, which is the same as "10[a]".
Example 4:

Input: "aabcaabcd"
Output: "2[aabc]d"
Explanation: "aabc" occurs twice, so one answer can be "2[aabc]d".
Example 5:

Input: "abbbabbbcabbbabbbc"
Output: "2[2[abbb]c]"
Explanation: "abbbabbbc" occurs twice, but "abbbabbbc" can also be encoded to "2[abbb]c", so one answer can be "2[2[abbb]c]".
 
This question lets us compress the string, enclose the same string in square brackets, and then add the number of occurrences in front. It feels like a rather difficult question. Refer to the post of the great god on the Internet to figure out what to do. This question should still be done with DP. We create a two-dimensional DP array, where dp[i][j] represents the abbreviated form of the string whose s is in the range of [i, j] (if the length of the abbreviated form is greater than the substring, then the substring is still retained) , Then if the length of the s string is n, the final result we need is stored in dp[0][n-1], and then we need to traverse all substrings of s, for any segment of substring [i, j], we \\ we split into two segments at any position k in the middle, and compare the total length of dp[i][k] plus dp[k+1][j] and dp[i][j] Length, assign a string with a smaller length to dp[i][j], and then what we have to do is to take the substring t in the range of [i, j] from s and merge it. The method of merging is that we add a t after the extracted string t, and then find the second starting position of the substring t in it. If the second starting position is less than the length of t, it means t Contains repeated strings, for example, such as t = "abab", then t + t = "abababab", we find the second t appears in the position 2, which is less than the length of t 4, indicating that there is Repeats, the number of repetitions is t.size()/pos = 2, then we need to put the repeated place in square brackets. Note that this substring cannot be placed directly in the square brackets, but from dp Take out the string at the corresponding position, because the repeated part may have been written in abbreviated form, such as example 5 in the title. Let's look at another example, if t = "abc", then t+t = "abcabc", We find the position of the second t appearing in it is 3, which is equal to the length of t 3, indicating that there is no repetition in t, then the replacement is still t. Then we compare the length of the string in the replace and dp[i][j] we get, and assign the smaller length to dp[i][j]. The time complexity is O(n 3 ) and the space is complicated. The degree is O(n 2 ), see the code as follows: