Create 269 - Alien Dictionary
https://zhuhan0.blogspot.com/2017/06/leetcode-269-alien-dictionary.html
There is a new alien language which uses the latin alphabet. However, the order among letters are unknown to you. You receive a list of non-empty words from the dictionary, where words are sorted lexicographically by the rules of this new language. Derive the order of letters in this language.
Example 1:
Given the following words in dictionary,
[
  "wrt",
  "wrf",
  "er",
  "ett",
  "rftt"
]
The correct order is: "wertf".
Example 2:
Given the following words in dictionary,
[
  "z",
  "x"
]
The correct order is: "zx".
Example 3:
Given the following words in dictionary,
[
  "z",
  "x",
  "z"
]
The order is invalid, so return "".
Note:
You may assume all letters are in lowercase.
You may assume that if a is a prefix of b, then a must appear before b in the given dictionary.
If the order is invalid, return an empty string.
There may be multiple valid order of letters, return any one of them is fine.
Topological sort:
Build graph: 
a map of character -> set of character.
Also get in-degrees for each character. In-degrees will be a map of character -> integer.
Topological sort:
Loop through in-degrees. Offer the characters with in-degree of 0 to queue.
While queue is not empty:
Poll from queue. Append to character to result string.
Decrease the in-degree of polled character's children by 1.
If any child's in-degree decreases to 0, offer it to queue.
At last, if result string's length is less than the number of vertices, that means there is a cycle in my graph. The order is invalid.
Say the number of characters in the dictionary (including duplicates) is n. Building the graph takes O(n). Topological sort takes O(V + E). V <= n. E also can't be larger than n. So the overall time complexity is O(n).