Create Print combinations of distinct numbers which add up to give sum N
Given a positive integer N, the task is to find out all the combinations of positive integers that add upto the given integer N. The program should print only combinations, not permutations and all the integers in a combination must be distinct. For example, for input 3, either 1, 2 or 2, 1 should be printed and 1, 1, 1 must not be printed as the integers are not distinct.

Examples:

Input: N = 3
Output:
1 2
3

Input: N = 7
Output:
1 2 4
1 6
2 5
3 4
7

Recommended: Please try your approach on {IDE} first, before moving on to the solution.
Approach: The approach is an extension of the approach discussed here. The idea used to get all the distinct element is that first we find all the elements that add up to give sum N. Then we iterate over each of the elements and store the elements into the set. Storing the elements into set would remove all the duplicate elements, and after that we add up the sum of the elements of the set and check whether it is equal to N or not.




Below is the implementation of the above approach: