Create Given a string, print all possible palindromic partitions
Given a string, find all possible palindromic partitions of given string.

Example:
AllPalindromPArtition

Recommended: Please try your approach on {IDE} first, before moving on to the solution.
Note that this problem is different from Palindrome Partitioning Problem, there the task was to find the partitioning with minimum cuts in input string. Here we need to print all possible partitions.

The idea is to go through every substring starting from first character, check if it is palindrome. If yes, then add the substring to solution and recur for remaining part. Below is complete algorithm.

Below is the implementation of above idea.